syncbot/requirements.txt
certifi==2026.1.4 ; python_version >= "3.11" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.11" and python_version < "4.0"
cryptography==46.0.5 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.1.1 ; python_version >= "3.11" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
idna==3.11 ; python_version >= "3.11" and python_version < "4.0"
pillow-heif==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==12.1.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.11" and python_version < "4.0"
pymysql==1.1.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.2.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.11" and python_version < "4.0"
slack-bolt==1.27.0 ; python_version >= "3.11" and python_version < "4.0"
slack-sdk==3.40.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==1.4.54 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.6.3 ; python_version >= "3.11" and python_version < "4.0"
